% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ifelser.R
\name{create_call}
\alias{create_call}
\title{Create ifelse call}
\usage{
create_call(lst)
}
\arguments{
\item{lst}{a chained ifelse statement using ifelser structure

This function needs to be rewritten}
}
\description{
Create ifelse call
}

