% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ifelser.R
\name{if_false}
\alias{if_false}
\title{Function that gets what to do if no}
\usage{
if_false(lst, no = NULL)
}
\arguments{
\item{lst}{a chained ifelse statement using ifelser structure}

\item{no}{return values for false elements of test}
}
\description{
Function that gets what to do if no
}

