% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ifelser.R
\name{if_false_then}
\alias{if_false_then}
\title{A more readable function to continue test_if statements}
\usage{
if_false_then(lst)
}
\description{
A more readable function to continue test_if statements
}

