% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ifelser.R
\name{if_true}
\alias{if_true}
\title{Function that gets what to do if yes}
\usage{
if_true(lst, yes = NULL)
}
\arguments{
\item{lst}{a chained ifelse statement using ifelser structure}

\item{yes}{return values for true elements of test}
}
\description{
Function that gets what to do if yes
}

