% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ifelser.R
\name{test_if.default}
\alias{test_if.default}
\title{Método do test_if quando o argumento é uma lista}
\usage{
\method{test_if}{default}(lst, test)
}
\arguments{
\item{lst}{a chained ifelse statement using ifelser structure}

\item{test}{an object which can be coerced to logical mode.}
}
\description{
Método do test_if quando o argumento é uma lista
}

