% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ifelser.R
\name{test_if.logical}
\alias{test_if.logical}
\title{Método do test_if quiando o argumento é um teste.}
\usage{
\method{test_if}{logical}(test)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}
}
\description{
Método do test_if quiando o argumento é um teste.
}

